src/test/java/org/omega/omegapoisk/service/AnimeContentServiceTest.java
package org.omega.omegapoisk.service;

import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.omega.omegapoisk.entity.content.Anime;
import org.omega.omegapoisk.service.content.AnimeContentService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.DynamicPropertyRegistry;
import org.springframework.test.context.DynamicPropertySource;
import org.springframework.transaction.annotation.Transactional;
import org.testcontainers.containers.PostgreSQLContainer;
import org.testcontainers.junit.jupiter.Container;
import org.testcontainers.junit.jupiter.Testcontainers;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
@Testcontainers
@Transactional
class AnimeContentServiceTest {
    @Container
    static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>(
            "postgres:latest"
    );

    @Autowired
    AnimeContentService animeContentService;

    @BeforeAll
    static void beforeAll() {
        postgres.start();
    }

    @AfterAll
    static void afterAll() {
        postgres.stop();
    }

    @DynamicPropertySource
    static void configureProperties(DynamicPropertyRegistry registry) {
        registry.add("spring.datasource.url", postgres::getJdbcUrl);
        registry.add("spring.datasource.username", postgres::getUsername);
        registry.add("spring.datasource.password", postgres::getPassword);
    }


    @Test
    void shouldGetAnimeById() {
        Anime createdAnime = animeContentService.create(new Anime(13));
        Anime retrievedAnime = animeContentService.getById(createdAnime.getId());

        assertThat(retrievedAnime).isNotNull();
        assertThat(retrievedAnime.getId()).isEqualTo(createdAnime.getId());
        assertThat(retrievedAnime.getSeriesNum()).isEqualTo(createdAnime.getSeriesNum());
    }


}